requirements.txt
coverage==4.2
mock==2.0.0
nose==1.3.7
pbr==1.10.0
python-redmine==1.5.1
PyYAML==3.11
requests==2.11.0
six==1.10.0
